Add datasource schema and test data; set ORM ddl-auto to update

+++ src/main/java/org/martynas/realestate_api/model/BuildingRecord.java
@@ -29,7 +29,7 @@ public class BuildingRecord {
     private int sizeInSquareMetres;
 
     @Column(name = "market_value")
-    private int marketValue;
+    private double marketValue;
 
     //    @Enumerated(EnumType.STRING)
     //    @Column(name = "property_type")
